+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0
+- Added animation groups
+- Added description, short description and title to animation DTO
+
 ## 0.1.0
 - First productive version of the server (Hooray!)
 - Featuring: